+++ src/wasm-stack.h
@@ -472,7 +472,7 @@ private:
   Function* func;
 };
 
-std::ostream& printStackIR(std::ostream& o, Module* module);
+std::ostream& printStackIR(std::ostream& o, Module* module, bool optimize);
 
 } // namespace wasm